% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiagnostics-package.r
\docType{package}
\name{netdiagnostics}
\alias{netdiagnostics}
\alias{netdiagnostics-package}
\title{netdiagnostics: a package for combining binary test results.}
\description{
netdiagnostics: a package for combining binary test results.
}
