% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_preformance.R
\docType{data}
\name{test_preformance}
\alias{test_preformance}
\title{Sample test performance statistics}
\format{A list of 2 tests:
\describe{
  \item{level 1}{test name}
  \item{level 2}{sensitivity and specificity, probability (vector)}
}}
\usage{
test_preformance
}
\description{
Sensitivity and specificity.
}
\keyword{datasets}
