% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_test_outcome_combination.R
\name{prob_test_outcome_combination}
\alias{prob_test_outcome_combination}
\title{prob_test_outcome_combination}
\usage{
prob_test_outcome_combination(prob, n_choose)
}
\arguments{
\item{n_choose}{}
}
\description{
prob_test_outcome_combination
}
